Rely on transaction state for commit timestamps value
Some type classes are defined as delegates of other types, so the arity
of the `serialize` method will be -1 as it delegates the implementation.
This causes errors for timestamp columns that are timezone aware. This
detail was added due to the conditional nature of the Spanner commit
timestamp value depending on the query type, dml or in a mutation. The
mutation logic is exposed by the connection, so we can delegate to that
instead of checking the arity of the method. This is a bit slower.